planner_prod/requirements.txt
aiogram==3.20.0
apscheduler==3.10.4
asyncpg==0.30.0
aiohttp==3.11.18
cryptography==45.0.0
geopy==2.4.1
openai==1.78.0
python-dotenv==1.1.0
redis==5.2.1
timezonefinder==6.5.9
yookassa==3.4.0
stripe==11.4.1
